requirements.txt
streamlit
Pillow
numpy
tensorflow
keras
opencv-python
scikit-learn
matplotlib
